Finished error back propagation and refactored services to be more granular
