+++ NeuralNetworks.Entities/IWorkerNeuron.cs
@@ -6,11 +6,6 @@
     public interface IWorkerNeuron : INeuron
     {
         /// <summary>
-        /// Gets or sets Back Propagated Error
-        /// </summary>
-        double BackPropagatedError { get; set; }
-
-        /// <summary>
         /// Gets or sets Output
         /// </summary>
         double Output { get; set; }
