+++ NeuralNetworks.Entities/Implementation/WorkerNeuron.cs
@@ -6,11 +6,6 @@
     public class WorkerNeuron : Neuron, INeuron, IWorkerNeuron
     {
         /// <summary>
-        /// Gets or sets Back Propagated Error
-        /// </summary>
-        public double BackPropagatedError { get; set; }
-
-        /// <summary>
         /// Gets or sets Output
         /// </summary>
         public double Output { get; set; }
